# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Centoire API",
    "version": "0.1.0",
    "description": "Centoire community platform API"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ]
      },
...
      "ViewerState": {
        "type": "object",
        "properties": {
          "voted": {
            "anyOf": [
              {
                "type": "number",
                "enum": [
                  1
                ]
              },
              {
                "type": "number",
                "enum": [
                  -1
                ]
              },
              {
                "nullable": true
              }
            ]
          },
          "bookmarked": {
            "type": "boolean"
          }
        },
        "required": [
          "voted",
          "bookmarked"
        ]
      },
      "PostCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "origin": {
            "type": "string",
            "enum": [
              "native",
              "aggregated"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "removed"
            ]
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          },
          "externalUrl": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string"
              },
              "handle": {
                "type": "string",
                "nullable": true
              },
              "displayName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "handle",
              "displayName",
              "avatarUrl"
            ]
          },
          "source": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "siteUrl": {
                "type": "string"
              },
              "faviconUrl": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "siteUrl",
              "faviconUrl"
            ]
          },
          "circle": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "slug"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "slug"
              ]
            }
          },
          "upvoteCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          },
          "bookmarkCount": {
            "type": "number"
          },
          "readTimeMinutes": {
            "type": "number"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string"
          },
          "viewer": {
            "$ref": "#/components/schemas/ViewerState"
          }
        },
        "required": [
          "id",
          "slug",
          "origin",
          "status",
          "title",
          "excerpt",
          "coverImageUrl",
          "externalUrl",
          "author",
          "source",
          "circle",
          "tags",
          "upvoteCount",
          "commentCount",
          "bookmarkCount",
          "readTimeMinutes",
          "publishedAt",
          "updatedAt",
          "viewer"
        ]
      },
      "PostDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PostCard"
          },
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "object",
                "nullable": true,
                "additionalProperties": {
                  "nullable": true
                }
              },
              "viewCount": {
                "type": "number"
              }
            },
            "required": [
              "content",
              "viewCount"
            ]
          }
        ]
      },
      "CreatePostInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "content": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 5
          },
          "circleId": {
            "type": "string",
            "nullable": true
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdatePostInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "content": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 5
          },
          "circleId": {
            "type": "string",
            "nullable": true
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "style",
              "craft",
              "business",
              "culture"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "postCount": {
            "type": "number"
          },
          "followerCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "category",
          "description",
          "postCount",
          "followerCount"
        ]
      },
...
      "VoteInput": {
        "type": "object",
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number",
                "enum": [
                  1
                ]
              },
              {
                "type": "number",
                "enum": [
                  -1
                ]
              }
            ]
          }
        },
        "required": [
          "value"
        ]
      },
      "BookmarkInput": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "deleted"
            ]
          },
          "depth": {
            "type": "number"
          },
          "upvoteCount": {
            "type": "number"
          },
          "replyCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "author": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string"
              },
              "handle": {
                "type": "string",
                "nullable": true
              },
              "displayName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              },
              "reputation": {
                "type": "number"
              }
            },
            "required": [
              "id",
              "handle",
              "displayName",
              "avatarUrl",
              "reputation"
            ]
          },
          "viewerVoted": {
            "type": "boolean"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        },
        "required": [
          "id",
          "content",
          "status",
          "depth",
          "upvoteCount",
          "replyCount",
          "createdAt",
          "author",
          "viewerVoted",
          "replies"
        ]
      },
      "CreateCommentInput": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "parentId": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateCommentInput": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "content"
        ]
      },
      "BookmarkFolders": {
        "type": "object",
        "properties": {
          "folders": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "count": {
                  "type": "number"
                }
              },
              "required": [
                "id",
                "name",
                "count"
              ]
            }
          },
          "unfiledCount": {
            "type": "number"
          }
        },
        "required": [
          "folders",
          "unfiledCount"
        ]
      },
      "CreateFolderInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "operationId": "listTags",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "style",
                "craft",
                "business",
                "culture"
              ]
            },
            "required": false,
            "name": "category",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "All tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCard"
                }
              }
            }
          },
          "403": {
            "description": "Email not verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/{id}": {
      "patch": {
        "tags": [
          "posts"
        ],
        "operationId": "updatePost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetail"
                }
              }
            }
          },
          "403": {
            "description": "Not your post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "operationId": "deletePost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Post removed"
          }
        }
      }
    },
    "/posts/{id}/publish": {
      "post": {
        "tags": [
          "posts"
        ],
        "operationId": "publishPost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Published post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCard"
                }
              }
            }
          },
          "422": {
            "description": "Post is missing title or content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/{id}/vote": {
      "put": {
        "tags": [
          "engagement"
        ],
        "operationId": "votePost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Vote recorded"
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "operationId": "unvotePost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Vote removed"
          }
        }
      }
    },
    "/comments/{id}/vote": {
      "put": {
        "tags": [
          "engagement"
        ],
        "operationId": "voteComment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Vote recorded"
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "operationId": "unvoteComment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Vote removed"
          }
        }
      }
    },
    "/posts/{id}/bookmark": {
      "put": {
        "tags": [
          "engagement"
        ],
        "operationId": "bookmarkPost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Bookmarked"
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "operationId": "unbookmarkPost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Bookmark removed"
          }
        }
      }
    },
    "/bookmarks": {
      "get": {
        "tags": [
          "engagement"
        ],
        "operationId": "listBookmarks",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "folderId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmarked posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostCard"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookmark-folders": {
      "get": {
        "tags": [
          "engagement"
        ],
        "operationId": "listBookmarkFolders",
        "responses": {
          "200": {
            "description": "Your folders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkFolders"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "engagement"
        ],
        "operationId": "createBookmarkFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "count": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "count"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Duplicate folder name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bookmark-folders/{id}": {
      "delete": {
        "tags": [
          "engagement"
        ],
        "operationId": "deleteBookmarkFolder",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Folder deleted; bookmarks unfiled"
          }
        }
      }
    },
    "/posts/{id}/comments": {
      "get": {
        "tags": [
          "engagement"
        ],
        "operationId": "listComments",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "engagement"
        ],
        "operationId": "createComment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Email not verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/comments/{id}": {
      "patch": {
        "tags": [
          "engagement"
        ],
        "operationId": "updateComment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Comment updated"
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "operationId": "deleteComment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted"
          }
        }
      }
    },
...
  }
}